make aside filter a component
