+++ src/components/ExperienceSearch/Filter/Filter.module.css
@@ -1,0 +1,37 @@
+@value main-yellow from 'common/variables.module.css';
+
+.frontButton, .rearButton {
+  display: inline-block;
+  width: 50%;
+  height: 32px;
+  line-height: 32px;
+  text-align: center;
+  border: 1px solid main-yellow;
+
+  &.toggle {
+    background-color: main-yellow;
+  }
+}
+
+.frontButton {
+  border-radius: 5px 0 0 5px;
+}
+
+.rearButton {
+  border-radius: 0 5px 5px 0;
+}
+
+.splitter {
+  margin-bottom: 32px;
+  margin-top: 32px;
+
+  @media (max-width: 850px) {
+    display: none;
+  }
+}
+
+.fliters {
+  @media (max-width: 850px) {
+    display: none;
+  }
+}
