+++ src/components/ExperienceSearch/index.js
@@ -4,7 +4,6 @@ import Helmet from 'react-helmet';
 import InfiniteScroll from 'react-infinite-scroller';
 import ReactGA from 'react-ga';
 
-import Checkbox from 'common/form/Checkbox';
 import Loader from 'common/Loader';
 import { Section, Wrapper } from 'common/base';
 
@@ -12,6 +11,7 @@ import styles from './ExperienceSearch.module.css';
 import Searchbar from './Searchbar';
 import ExperienceBlock from './ExperienceBlock';
 import WorkingHourBlock from './WorkingHourBlock';
+import Filter from './Filter';
 import { Banner1, Banner2 } from './Banners';
 import { fetchExperiences } from '../../actions/experienceSearch';
 
@@ -20,11 +20,9 @@ import getScale from '../../utils/numberUtils';
 
 import { GA_CATEGORY, GA_ACTION } from '../../constants/gaConstants';
 
-const SORT = { CREATED_AT: 'created_at', POPULARITY: 'popularity' };
-const SEARCH_TYPE = {
-  INTERVIEW: 'interview',
-  WORK: 'work',
-  SALARY: 'salary',
+const SORT = {
+  CREATED_AT: 'created_at',
+  POPULARITY: 'popularity',
 };
 
 class ExperienceSearch extends Component {
@@ -143,42 +141,11 @@ class ExperienceSearch extends Component {
         <Wrapper size="l">
           <div className={styles.container}>
             <aside className={styles.aside}>
-              <section>
-                <button
-                  className={data.sort === SORT.CREATED_AT
-                    ? `${styles.frontButton} ${styles.toggle}`
-                    : styles.frontButton}
-                  onClick={this.fetchExperiencesWithSort} value={SORT.CREATED_AT}
-                >
-                  最新
-                </button>
-                <button
-                  className={data.sort === SORT.POPULARITY
-                    ? `${styles.rearButton} ${styles.toggle}`
-                    : styles.rearButton}
-                  onClick={this.fetchExperiencesWithSort} value={SORT.POPULARITY}
-                >
-                  熱門
-                </button>
-              </section>
-              <hr className={styles.splitter} />
-              <div className={styles.fliters}>
-                {
-                  [
-                    { label: '面試經驗', value: SEARCH_TYPE.INTERVIEW },
-                    { label: '工作經驗', value: SEARCH_TYPE.WORK },
-                    { label: '薪資工時', value: SEARCH_TYPE.SALARY },
-                  ].map(o => (
-                    <Checkbox
-                      key={o.value} id={`searchType-${o.value}`}
-                      label={o.label} value={o.value}
-                      disabled={o.value === 'salary' && !data.searchQuery}
-                      onChange={this.setSearchType}
-                      checked={data[o.value]}
-                    />
-                  ))
-                }
-              </div>
+              <Filter
+                data={data}
+                fetchExperiencesWithSort={this.fetchExperiencesWithSort}
+                setSearchType={this.setSearchType}
+              />
               <Banner1 />
             </aside>
